Got LlamaXML building under Visual Studio .NET 2003 on Windows.

+++ LifeBalance/branches/WindowsMobile/LlamaXML/Common/LlamaXML/XMLReader.cpp
@@ -644,7 +644,7 @@ namespace LlamaXML {
 		}
 		size_t remainingCapacity = mInputBuffer + kInputBufferCount - mInputEnd;
 		if (remainingCapacity > 0) {
-			mInputEnd += mInput.ReadUpTo(mInputEnd, remainingCapacity);
+			mInputEnd += mInput.ReadUpTo(mInputEnd, uint32_t(remainingCapacity));
 		}
 	}
